+++ src/mjolnir/hierarchybuilder.cc
@@ -318,7 +318,7 @@ void FormTilesInNewLevel(GraphReader& reader,
       // encoded shape plus way Id.
       auto edgeinfo = tile->edgeinfo(idx);
       std::string encoded_shape = edgeinfo.encoded_shape();
-      uint64_t w = hasher(encoded_shape + std::to_string(edgeinfo.wayid()));
+      uint32_t w = hasher(encoded_shape + std::to_string(edgeinfo.wayid()));
       uint32_t edge_info_offset =
           tilebuilder->AddEdgeInfo(w, nodea, nodeb, edgeinfo.wayid(), edgeinfo.mean_elevation(),
                                    edgeinfo.bike_network(), edgeinfo.speed_limit(), encoded_shape,
